fix: correct scroll position persistence respect de zoom text

+++ app/src/main/java/com/juguito/juguitoreader/data/local/entity/ReadingProgressEntity.kt
@@ -9,7 +9,7 @@ import com.juguito.juguitoreader.domain.enums.SyncStatus
 data class ReadingProgressEntity (
     @ColumnInfo(name = "book_id") @PrimaryKey val bookId: Int,
     @ColumnInfo(name = "last_chapter_index") val lastChapterIndex: Int,
-    @ColumnInfo(name = "scroll_position") val scrollPosition: Int,
+    @ColumnInfo(name = "scroll_position") val scrollPosition: Float,
     @ColumnInfo(name = "last_read_at") val lastReadAt: Long,
     @ColumnInfo(name = "sync_status") val syncStatus: SyncStatus = SyncStatus.PENDING_CREATE
     )
