+++ app/src/main/java/com/juguito/juguitoreader/domain/model/ReadingProgress.kt
@@ -3,14 +3,14 @@ package com.juguito.juguitoreader.domain.model
 class ReadingProgress (
     val bookId: Int,
     val lastChapterIndex: Int,
-    val scrollPosition: Int,
+    val scrollPosition: Float,
     val lastReadAt: Long
 )
 
 fun ReadingProgress.copy(
     bookId: Int = this.bookId,
     lastChapterIndex: Int = this.lastChapterIndex,
-    scrollPosition: Int = this.scrollPosition,
+    scrollPosition: Float = this.scrollPosition,
     lastReadAt: Long = this.lastReadAt
 ): ReadingProgress {
     return ReadingProgress(
