+++ app/src/main/java/com/juguito/juguitoreader/ui/reader/ReaderEvent.kt
@@ -7,6 +7,6 @@ sealed interface ReaderEvent {
     data class OnChapterSelected(val index: Int): ReaderEvent
     data class OnTextZoomChanged(val zoom: Int): ReaderEvent
     data class OnThemeChanged(val theme: ReaderTheme): ReaderEvent
-    data class OnScrollPositionChanged(val scrollPosition: Int): ReaderEvent
+    data class OnScrollPositionChanged(val scrollPosition: Float): ReaderEvent
     data class OnTimeRemainingChanged(val minutes: Int) : ReaderEvent
 }
